backend/requirements.txt
# FastAPI
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# HTTP
requests>=2.31.0
httpx>=0.25.0
aiohttp>=3.9.0

# AI/ML - SAM 本地分割
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
pillow>=10.0.0
numpy>=1.24.0

# 数据库
sqlalchemy>=2.0.0
pymysql>=1.1.0

# 安全认证
bcrypt>=4.0.0
PyJWT>=2.8.0

# 其他
python-dotenv>=1.0.0
pydantic>=2.0.0
aiofiles>=23.0.0
